Refactor models into their own package
